# openapi.json
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /support/:
    get:
      operationId: support_tickets_list
      description: Retrieve all support tickets belonging to the authenticated user.
      summary: List my support tickets
      tags:
      - Support
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicketListResponse'
          description: ''
    post:
      operationId: support_ticket_create
      description: Create a new support ticket with an initial message for the authenticated
        user.
      summary: Create a support ticket
      tags:
      - Support
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateTicket'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateTicket'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTicketResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /support/{ticket_id}/:
    get:
      operationId: support_ticket_retrieve
      description: Retrieve a single support ticket with its message thread for the
        authenticated user.
      summary: Get support ticket detail
      parameters:
      - in: path
        name: ticket_id
        schema:
          type: integer
        required: true
      tags:
      - Support
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    post:
      operationId: support_ticket_add_message
      description: Append a message to an existing support ticket owned by the authenticated
        user.
      summary: Add a message to a support ticket
      parameters:
      - in: path
        name: ticket_id
        schema:
          type: integer
        required: true
      tags:
      - Support
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AddMessage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AddMessage'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddMessageResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
...
    AddMessage:
      type: object
      properties:
        message:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
          writeOnly: true
      required:
      - message
    AddMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          $ref: '#/components/schemas/SupportMessage'
      required:
      - message
      - success
...
    CreateTicket:
      type: object
      properties:
        subject:
          type: string
          maxLength: 255
        description:
          type: string
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        images:
          type: array
          items:
            type: string
            format: uri
          writeOnly: true
      required:
      - description
      - subject
    CreateTicketResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        ticket:
          $ref: '#/components/schemas/SupportTicket'
      required:
      - message
      - success
      - ticket
...
    PriorityEnum:
      enum:
      - low
      - medium
      - high
      - urgent
      type: string
      description: |-
        * `low` - Low
        * `medium` - Medium
        * `high` - High
        * `urgent` - Urgent
...
    SupportAttachment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        image:
          type: string
          format: uri
          nullable: true
          readOnly: true
        uploaded_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - id
      - image
      - uploaded_at
    SupportMessage:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        sender_name:
          type: string
          readOnly: true
        message:
          type: string
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/SupportAttachment'
          readOnly: true
      required:
      - attachments
      - created_at
      - id
      - message
      - sender_name
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        subject:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          $ref: '#/components/schemas/SupportTicketStatusEnum'
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        messages:
          type: array
          items:
            $ref: '#/components/schemas/SupportMessage'
          readOnly: true
      required:
      - created_at
      - description
      - id
      - messages
      - subject
      - updated_at
    SupportTicketListResponse:
      type: object
      properties:
        count:
          type: integer
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/SupportTicket'
      required:
      - count
      - tickets
    SupportTicketStatusEnum:
      enum:
      - open
      - in_progress
      - resolved
      - closed
      type: string
      description: |-
        * `open` - Open
        * `in_progress` - In Progress
        * `resolved` - Resolved
        * `closed` - Closed
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT